requirements.txt
dal
django
django-polymorphic==2.1.2
django-phonenumber-field==4.0.0
phonenumbers==8.11.5
whitenoise==5.0.1